スキルシートを実装改善
