+++ app/models/personal_skill.server.ts
@@ -1,6 +1,10 @@
 import { PersonalSkillList, Term, User } from "@prisma/client";
 import { prisma } from "~/db.server";
 
+export const getSkillCategories = async () => {
+  return await prisma.skillCategory.findMany({});
+};
+
 export const getPersonalSkillList = async (user: User, term: Term) => {
   const skillList = await prisma.personalSkillList.findUnique({
     where: {
@@ -58,6 +62,7 @@ export const copySkills = async (
       personalSkillListId: to.id,
       skillId: skill.skillId,
       experienceYear: skill.experienceYear,
+      skillCategoryId: skill.skillCategoryId,
     })),
   });
 };
@@ -66,6 +71,7 @@ export const upsertPersonalSkill = async (args: {
   personalSkillListId: number;
   skillId: number;
   experienceYear: number;
+  skillCategoryId: number;
 }) => {
   return await prisma.personalSkill.upsert({
     where: {
@@ -81,26 +87,7 @@ export const upsertPersonalSkill = async (args: {
       personalSkillListId: args.personalSkillListId,
       skillId: args.skillId,
       experienceYear: args.experienceYear,
-    },
-  });
-};
-export const insertPersonalSkill = async (args: {
-  userId: number;
-  termId: number;
-}) => {
-  const skillList = await prisma.personalSkillList.findUnique({
-    where: {
-      userId_termId: {
-        userId: args.userId,
-        termId: args.termId,
-      },
-    },
-  });
-  return await prisma.personalSkill.create({
-    data: {
-      personalSkillListId: skillList!.id,
-      skillId: 1,
-      experienceYear: 0,
+      skillCategoryId: args.skillCategoryId,
     },
   });
 };
